Configure routers based on use case
This PR makes it possible to configure which routers get mounted based on the use case. The main reason for this at the moment is not to mount the template model API to the space weather deployment.

It does this by passing the `domain` set in the dockerfile through to `startup.sh` then setting it as an environment variable so it can get read by the WSGI script inside python. There's probably a  better way to do this - @kkaris can you advise?